Move article_doi and set_article_doi to AbstractNode model.
